Add condition to not ad default "size type" in "px" or "%" is present
For the Percent or Pixels, changed the situation to allow "custom sizing" and not forcing to pixels or percents, per Admin setting.
This way, if a custom-size player (ex: 300px by 100px) is necessary, the size will not be warped into percents (ex previously: 300px% by 100px%, currently: 300px by 100px)

The size-types is added if it is not present only, leaving the user to decide which size should be used.